Move legacy send scene logic to the right place
