+++ src/__tests__/reducers/sendConfirmationReducer.test.ts
@@ -3,8 +3,7 @@ import { EdgeTransaction } from 'edge-core-js'
 import { cloneDeep } from 'lodash'
 
 import s from '../../locales/strings'
-import { initialState } from '../../modules/UI/scenes/SendConfirmation/selectors'
-import { sendConfirmation } from '../../reducers/scenes/SendConfirmationReducer'
+import { initialState, sendConfirmation } from '../../reducers/scenes/SendConfirmationReducer'
 
 describe('sendConfirmation reducer', () => {
   test('initialState', () => {
